Backend/requirements.txt
fastapi
uvicorn
numpy
matplotlib
scipy
Pillow
cython
scikit-image
tensorflow>=1.3.0
keras>=2.0.8
opencv-python
h5py
imgaug
IPython[all]